requirements.txt
beautifulsoup4==4.11.1
dnspython==2.1.0
gspread==3.6.0
protobuf==4.21.1
pymongo==3.11.2
python-dotenv==0.20.0
requests==2.27.1
selenium==3.141.0
webdriver_manager==3.3.0
